+++ src/Network/Network/NetworkVirtualAppliance/NewNetworkVirtualApplianceCommand.cs
@@ -144,6 +144,13 @@ namespace Microsoft.Azure.Commands.Network
         [ValidateNotNullOrEmpty]
         public PSVirtualApplianceInternetIngressIpsProperties[] InternetIngressIp { get; set; }
 
+        [Parameter(
+            Mandatory = false,
+            ValueFromPipelineByPropertyName = true,
+            HelpMessage = "Network Profile of the Virtual Appliance.")]
+        [ValidateNotNullOrEmpty]
+        public PSVirtualApplianceNetworkProfile NetworkProfile { get; set; }
+
         public override void Execute()
         {
             base.Execute();
@@ -192,6 +199,11 @@ namespace Microsoft.Azure.Commands.Network
             if (InternetIngressIp != null)
             {
                 networkVirtualAppliance.InternetIngressPublicIps = InternetIngressIp;
+            }
+
+            if (NetworkProfile != null)
+            {
+                networkVirtualAppliance.NetworkProfile = NetworkProfile;
             }
 
             var networkVirtualApplianceModel = NetworkResourceManagerProfile.Mapper.Map<MNM.NetworkVirtualAppliance>(networkVirtualAppliance);